+++ src/Account/AccountInterface.java
@@ -6,26 +6,26 @@ public interface AccountInterface {
      * Deposits amount into the current account
      * @param amount integer
      */
-    void deposit(float amount);
+    void deposit(double amount);
 
     /**
      * Withdraws amount from current account balance if balance is greater than the amount requested.
      * Otherwise does nothing.
      * @param amount integer
      */
-    void withdraw(float amount);
+    void withdraw(double amount);
 
     /**
      * Returns boolean if the account has sufficient funds for the amount requested.
      * @param amount
      * @return boolean
      */
-    boolean isSufficientFunds(float amount);
+    boolean isSufficientFunds(double amount);
 
     /**
      * Returns the balance of the account
      * @return balance
      */
-    float getBalance();
+    double getBalance();
 
 }
